requirements.txt
asgiref==3.3.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-autocomplete-light==3.8.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
flake8==3.8.4
idna==2.10
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
